requirements.txt
# Core RAG System Dependencies
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.13
langgraph==0.0.26
langsmith==0.0.69

# LLM and Embeddings
openai==1.6.1
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.19.4

# Vector Databases
qdrant-client==1.7.0
chromadb==0.4.18

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
unstructured==0.11.6

# Text Processing
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2

# Evaluation Metrics
bert-score==0.3.13
rouge-score==0.1.2
sacrebleu==2.3.1
evaluate==0.4.1

# Web Framework and Dashboard
streamlit==1.29.0
dash==2.16.1
plotly==5.17.0
gradio==4.8.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Database and Storage
sqlalchemy==2.0.23
sqlite3

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and Monitoring
loguru==0.7.2
wandb==0.16.1

# HTTP and API
requests==2.31.0
httpx==0.25.2
fastapi==0.104.1
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Async Support
asyncio
aiofiles==23.2.1